fix: 엔티티에 맞게 반환 타입을 변경

+++ src/main/java/aegis/server/domain/payment/repository/TransactionRepository.java
@@ -4,10 +4,11 @@ import aegis.server.domain.payment.domain.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface TransactionRepository extends JpaRepository<Transaction, Long> {
 
     @Query("SELECT SUM(t.amount) FROM Transaction t WHERE t.depositorName = :depositorName")
-    Optional<Long> sumAmountByDepositorName(String depositorName);
+    Optional<BigDecimal> sumAmountByDepositorName(String depositorName);
 }
